Fixed the bug, debug code still in
The Gutenberg files with the headers removed will
have several empty lines. If an empty line is sent
to the parser, it will segfault.

The bug is fixed by first checking that the line
stripped of \n is not zero length and then a new
variable to keep track of parsed lines is added as the
enumerate indexing doesn't work anymore.